requirements.txt
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
Flask==2.2.2
fonttools==4.38.0
fpdf2==2.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
Werkzeug==2.2.2
